Addressed review comments

+++ include/motor-control/firmware/motor_control_hardware.h
@@ -20,8 +20,8 @@ bool motor_hardware_set_dac_value(void* dac_handle, uint32_t channel,
                                   uint32_t data_algn, uint32_t val);
 bool motor_hardware_start_pwm(void* tim_handle, uint32_t channel);
 bool motor_hardware_stop_pwm(void* tim_handle, uint32_t channel);
-uint32_t motor_hardware_encoder_pulse_count(void* encoder_handle);
-uint32_t motor_hardware_encoder_pulse_count_with_overflow(void* encoder_handle, int8_t *overflows);
+uint16_t motor_hardware_encoder_pulse_count(void* encoder_handle);
+uint16_t motor_hardware_encoder_pulse_count_with_overflow(void* encoder_handle, int8_t *overflows);
 void motor_hardware_reset_encoder_count(void* encoder_handle, uint16_t reset_value);
 uint16_t motor_hardware_get_stopwatch_pulses(void* stopwatch_handle, uint8_t clear);
 #ifdef __cplusplus
